Rename Local and Session Storage
